Create House Robber iv
Here in this question j gets assigned as the largest element in the array [in the worst case that one could be our answer].
After finding the mid we are checking that can we have the answer as mid, for that we need to include all the elements lesser than or equal to mid but not greater than mid, followed by this we are checking that the length of the subsequence we have formed is greater than or equal to k or not.
If yes then we will compress the range from the right and that could be one of the possible answers. So we will assign that as the answer.